move tests to cohort service
