+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppselectronicmonitoringcreateanorderapi/service/OrderService.kt
@@ -101,7 +101,7 @@ class OrderService(
     notifyingOrganisation: String?,
   ): Boolean {
     val userCohort = userCohortService.getUserCohort(token)
-    return userCohortService.matchesNofifyingOrg(userCohort, notifyingOrganisation)
+    return userCohortService.matchesNofifyingOrg(userCohort.cohort, notifyingOrganisation)
   }
 
   fun createVersion(orderId: UUID, token: JwtAuthenticationToken, versionType: RequestType): Order {
